Implement BellNotificationUI for order alerts and refactor related UI components
