+++ Assets/Scripts/UI/BellNotificationUI.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 59bb0748a8fe02c4c804e3e4cb93a7f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
